requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
Django==4.2.3
django-ckeditor==6.6.1
django-froala-editor==4.0.19
django-js-asset==2.1.0
django-quill-editor==0.1.40
django-tailwind==3.6.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==2.0.3